Trending, Upcoming, and Review compontents wireframed- functional

+++ api/migrations/04_reviews_table.py
@@ -5,7 +5,7 @@ steps = [
         CREATE TABLE reviews (
             id SERIAL PRIMARY KEY NOT NULL,
             username VARCHAR(50) NOT NULL REFERENCES accounts(username),
-            rating INTEGER NOT NULL,
+            rating FLOAT NOT NULL,
             review VARCHAR(500),
             movie_id INTEGER NOT NULL REFERENCES movies(tmdb_movie_id),
             created_at TIMESTAMP NOT NULL DEFAULT NOW()
