+++ api/queries/reviews.py
@@ -6,7 +6,7 @@ class Error(BaseModel):
 
 class MovieReviewIn(BaseModel):
     review: str
-    rating: int
+    rating: float
 
 
 class MovieReviewOut(BaseModel):
@@ -14,7 +14,7 @@ class MovieReviewOut(BaseModel):
     username: str
     movie_id: int
     review: str
-    rating: int
+    rating: float
 
 class MovieReviewRepo:
     def add_review(self, username: str, movie_id: int, review: MovieReviewIn) -> MovieReviewOut:
@@ -142,7 +142,7 @@ class MovieReviewRepo:
                     return True
                 else:
                     return False
-                    
+
     def update_review(self, id: int, review: MovieReviewIn) -> MovieReviewOut:
         with pool.connection() as conn:
             with conn.cursor() as db:
